Added check for symlink to ensure symlink is created

+++ backup.sh
@@ -15,7 +15,11 @@ mount -o rw,remount $PATHTOBACKUP
 rsync -aAXvz --exclude-from $EXCLUDEFILE --link-dest=$PATHTOBACKUP/current $SOURCE $PATHTOBACKUP/back-$date
 
 # Make new symlink for current
-rm $PATHTOBACKUP/current && ln -s back-$date $PATHTOBACKUP/current
+if [ -e $PATHTOBACKUP/current ]; then
+    rm $PATHTOBACKUP/current && ln -s back-$date $PATHTOBACKUP/current
+else
+    ln -s back-$date $PATHTOBACKUP/current
+fi
 
 # Write protect disk
 mount -o ro,remount $PATHTOBACKUP
